Move dynamic controller
